// multiplayer/mp-client/src/constants/OlympiaItemStats.generated.ts
export interface OlympiaItemStats {
    equipPos: number;
    effectType: number;
    effectValues: readonly [number, number, number, number, number, number];
    maxLifeSpan: number;
    weight: number;
    speed: number;
    levelLimit: number;
}